refactor(experiment_grid): move execute()/compose_experiment() to launch_grid.py
__init__.py is now a thin re-export; the logic lives in launch_grid.py like
any other module, keeping __init__.py easy to scan.
